\name{imprProbEst-package}
\alias{imprProbEst-package}
\alias{imprProbEst}
\docType{package}
\title{Minimum distance estimation in an imprecise probability model}
\description{
A minimum distance estimator is calculated for an imprecise probability model.
  The imprecise probability model consists of upper coherent previsions whose credal sets are given
  by a finite number of constraints on the expectations. The parameter set is finite. The estimator chooses
  that parameter such that the empirical measure lies next to the corresponding credal set with respect to
  the total variation norm.}
\details{
\tabular{ll}{
Package: \tab imprProbEst\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-10-23\cr
License: \tab LGPL-3\cr
LazyLoad: \tab yes\cr
}
library(imprProbEst
}
\author{
Robert Hable

Maintainer: Robert Hable <Robert.Hable@uni-bayreuth.de>
}
\references{
  Hable (2008) \emph{Data-Based Decisions under Complex Uncertainty}, Ph.D. thesis,
  LMU Munich, \emph{in preparation}

  Walley, P. (1991) \emph{Statistical reasoning with imprecise probabilities}. 
  Chapman & Hall, London.
}
\note{R programming support was given by Matthias Kohl}
%\seealso{}
%\examples{}
\keyword{ package }
